added new enpoint and refactor code endpoint
/complex-calc?query={query}
/add?leftOperation={leftOperand}&rightOperand={rightOperand}
/subtract?leftOperation={leftOperand}&rightOperand={rightOperand}
/multiply?leftOperation={leftOperand}&rightOperand={rightOperand}
/divide?leftOperation={leftOperand}&rightOperand={rightOperand}
/average?type={type}&set={1,2,3,4,5,6}

+++ JavaServerUnitTestExample/src/main/java/com/computation/rest/service/SimpleCalculatorService.java
@@ -2,12 +2,11 @@ package com.computation.rest.service;
 
 public interface SimpleCalculatorService {
 	
-	int add(int operand1, int operand2);
+	float add(float operand1, float operand2);
 
-	int subtract(int operand1, int operand2);
+	float subtract(float operand1, float operand2);
 
-	int multiply(int operand1, int operand2);
+	float multiply(float operand1, float operand2);
 
-	int divison(int operand1, int operand2);
-
+	float divison(float operand1, float operand2);
 }
